Initial Commit
I made it so that you don't have to login to get access to the website.
Also edited permissions so that users that are not logged-in can't see
or access certain .php files. Finally, I removed the guest account
functionality since logging in is no longer required. Session variables
are currently printed to the screen for testing.